feat: moved into `importer.rs` and did magic check
