+++ crates/asset-importer-rs-dae/tests/importer.rs
@@ -10,12 +10,20 @@ fn dummy_loader(_path: &Path) -> io::Result<Cursor<Vec<u8>>> {
     Ok(Cursor::new(Vec::new()))
 }
 
+fn collada_loader(_path: &Path) -> io::Result<Cursor<Vec<u8>>> {
+    Ok(Cursor::new(
+        b"<?xml version=\"1.0\"?><COLLADA xmlns=\"http://www.collada.org/2005/11/COLLADASchema\" version=\"1.4.1\">"
+            .to_vec(),
+    ))
+}
+
 #[test]
 fn test_dae_can_read_extension() {
     let importer = DaeImporter::new();
-    assert!(importer.can_read("model.dae", dummy_loader));
-    assert!(!importer.can_read("model.obj", dummy_loader));
-    assert!(!importer.can_read("model", dummy_loader));
+    assert!(importer.can_read("model.dae", collada_loader));
+    assert!(!importer.can_read("model.dae", dummy_loader));
+    assert!(!importer.can_read("model.obj", collada_loader));
+    assert!(!importer.can_read("model", collada_loader));
 }
 
 #[test]
